+++ NEWS.rst
@@ -8,7 +8,9 @@ snakeoil 0.11.1 (unreleased)
 API deprecations
 ~~~~~~~~~~~~~~~~
 
-* `snakeoil.contexts.patch`.  Use `unittest.mock.patch` instead.  Removal in `0.12.0`
+* `snakeoil.sequences.predicate_split`.  `snakeoil.iterables.partition` is
+  the iterable equivalent.  Use that instead.  Removal in `0.12.0`.
+* `snakeoil.contexts.patch`.  Use `unittest.mock.patch` instead.  Removal in `0.12.0`.
 
 Packaging
 ~~~~~~~~~~